# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AEVIUM AI Writing Assistant API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /chapters/{chapterId}/scenes:
    get:
      operationId: listScenes
      tags: [scenes]
      summary: List scenes in a chapter
      parameters:
        - name: chapterId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of scenes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Scene"
    post:
      operationId: createScene
      tags: [scenes]
      summary: Create a scene in a chapter
      parameters:
        - name: chapterId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateSceneBody"
      responses:
        "201":
          description: Created scene
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scene"
  /chapters/{chapterId}/scenes/{id}:
    get:
      operationId: getScene
      tags: [scenes]
      summary: Get a scene
      parameters:
        - name: chapterId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Scene
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scene"
    patch:
      operationId: updateScene
      tags: [scenes]
      summary: Update a scene
      parameters:
        - name: chapterId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateSceneBody"
      responses:
        "200":
          description: Updated scene
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scene"
    delete:
      operationId: deleteScene
      tags: [scenes]
      summary: Delete a scene
      parameters:
        - name: chapterId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Characters ────────────────────────────────────────────────────────────
...
  /projects/{projectId}/timeline:
    get:
      operationId: listTimelineEvents
      tags: [timelineEvents]
      summary: List timeline events in a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of timeline events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TimelineEvent"
    post:
      operationId: createTimelineEvent
      tags: [timelineEvents]
      summary: Create a timeline event
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTimelineEventBody"
      responses:
        "201":
          description: Created timeline event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimelineEvent"
  /projects/{projectId}/timeline/{id}:
    patch:
      operationId: updateTimelineEvent
      tags: [timelineEvents]
      summary: Update a timeline event
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateTimelineEventBody"
      responses:
        "200":
          description: Updated timeline event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TimelineEvent"
    delete:
      operationId: deleteTimelineEvent
      tags: [timelineEvents]
      summary: Delete a timeline event
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ── Style Guide ───────────────────────────────────────────────────────────
  /projects/{projectId}/style-guide:
    get:
      operationId: getStyleGuide
      tags: [styleGuides]
      summary: Get the style guide for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Style guide
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StyleGuide"
        "404":
          description: Not found
    put:
      operationId: upsertStyleGuide
      tags: [styleGuides]
      summary: Create or update the style guide for a project
      parameters:
        - name: projectId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpsertStyleGuideBody"
      responses:
        "200":
          description: Updated style guide
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StyleGuide"
components:
  schemas:
...
    Scene:
      type: object
      properties:
        id:
          type: integer
        chapterId:
          type: integer
        title:
          type: string
        position:
          type: integer
        content:
          type: ["string", "null"]
        summary:
          type: ["string", "null"]
        povCharacterId:
          type: ["integer", "null"]
        locationId:
          type: ["integer", "null"]
        timelinePosition:
          type: ["string", "null"]
        narrativeGoal:
          type: ["string", "null"]
        wordCount:
          type: integer
        status:
          type: string
          enum: [draft, in_review, ready, blocked, needs_rewrite, needs_continuity]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - chapterId
        - title
        - position
        - wordCount
        - status
        - createdAt
        - updatedAt
    CreateSceneBody:
      type: object
      properties:
        title:
          type: string
        position:
          type: integer
        content:
          type: ["string", "null"]
        summary:
          type: ["string", "null"]
        povCharacterId:
          type: ["integer", "null"]
        locationId:
          type: ["integer", "null"]
        timelinePosition:
          type: ["string", "null"]
        narrativeGoal:
          type: ["string", "null"]
        status:
          type: string
          enum: [draft, in_review, ready, blocked, needs_rewrite, needs_continuity]
      required:
        - title
    UpdateSceneBody:
      type: object
      properties:
        title:
          type: string
        position:
          type: integer
        content:
          type: ["string", "null"]
        summary:
          type: ["string", "null"]
        povCharacterId:
          type: ["integer", "null"]
        locationId:
          type: ["integer", "null"]
        timelinePosition:
          type: ["string", "null"]
        narrativeGoal:
          type: ["string", "null"]
        wordCount:
          type: integer
        status:
          type: string
          enum: [draft, in_review, ready, blocked, needs_rewrite, needs_continuity]

    # ── Character ─────────────────────────────────────────────────────────────
...
    TimelineEvent:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        bookId:
          type: ["integer", "null"]
        chapterId:
          type: ["integer", "null"]
        sceneId:
          type: ["integer", "null"]
        characterId:
          type: ["integer", "null"]
        eventType:
          type: string
          enum: [death, injury, travel, revelation, conflict, romance, political, worldbuilding, other]
        description:
          type: string
        dateLabel:
          type: ["string", "null"]
        orderIndex:
          type: integer
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - eventType
        - description
        - orderIndex
        - createdAt
    CreateTimelineEventBody:
      type: object
      properties:
        bookId:
          type: ["integer", "null"]
        chapterId:
          type: ["integer", "null"]
        sceneId:
          type: ["integer", "null"]
        characterId:
          type: ["integer", "null"]
        eventType:
          type: string
          enum: [death, injury, travel, revelation, conflict, romance, political, worldbuilding, other]
        description:
          type: string
        dateLabel:
          type: ["string", "null"]
        orderIndex:
          type: integer
      required:
        - eventType
        - description
        - orderIndex
    UpdateTimelineEventBody:
      type: object
      properties:
        eventType:
          type: string
          enum: [death, injury, travel, revelation, conflict, romance, political, worldbuilding, other]
        description:
          type: string
        dateLabel:
          type: ["string", "null"]
        orderIndex:
          type: integer
        characterId:
          type: ["integer", "null"]

    # ── StyleGuide ────────────────────────────────────────────────────────────
    StyleGuide:
      type: object
      properties:
        id:
          type: integer
        projectId:
          type: integer
        narrator:
          type: ["string", "null"]
          enum: [first_person, third_limited, third_omniscient, second_person, null]
        tense:
          type: ["string", "null"]
          enum: [past, present, null]
        povType:
          type: ["string", "null"]
          enum: [single, multiple, null]
        tone:
          type: ["string", "null"]
        sensorDetailLevel:
          type: ["string", "null"]
        violenceLevel:
          type: ["string", "null"]
        introspectionLevel:
          type: ["string", "null"]
        pacing:
          type: ["string", "null"]
          enum: [slow, medium, fast, null]
        forbiddenWords:
          type: ["string", "null"]
        frequentWords:
          type: ["string", "null"]
        dialogueRules:
          type: ["string", "null"]
        povRules:
          type: ["string", "null"]
        additionalNotes:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - projectId
        - createdAt
        - updatedAt
    UpsertStyleGuideBody:
      type: object
      properties:
        narrator:
          type: ["string", "null"]
          enum: [first_person, third_limited, third_omniscient, second_person, null]
        tense:
          type: ["string", "null"]
          enum: [past, present, null]
        povType:
          type: ["string", "null"]
          enum: [single, multiple, null]
        tone:
          type: ["string", "null"]
        sensorDetailLevel:
          type: ["string", "null"]
        violenceLevel:
          type: ["string", "null"]
        introspectionLevel:
          type: ["string", "null"]
        pacing:
          type: ["string", "null"]
          enum: [slow, medium, fast, null]
        forbiddenWords:
          type: ["string", "null"]
        frequentWords:
          type: ["string", "null"]
        dialogueRules:
          type: ["string", "null"]
        povRules:
          type: ["string", "null"]
        additionalNotes:
          type: ["string", "null"]